+++ bones/__init__.py
@@ -3,7 +3,6 @@ from . import base
 from . import boolean
 from . import color
 from . import date
-from . import file
 from . import numeric
 from . import record
 from . import relational
@@ -12,8 +11,5 @@ from . import spatial
 from . import string
 from . import text
 
-# from . import hierarchy
 # from . import email
 # from . import password
-# from . import treedir
-# from . import spatial